move fuzzing.h to be in the same directory as fuzzing.cpp.
